refactor(probe): split cli parser and report-io into submodules
